输出H5 app onLaunch 和 onShow 函数添加options
