+++ packages/core/src/platform/createApp.js
@@ -1,7 +1,7 @@
 import transferOptions from '../core/transferOptions'
 import mergeOptions from '../core/mergeOptions'
 import builtInKeysMap from './patch/builtInKeysMap'
-import { makeMap } from '../helper/utils'
+import { makeMap, getUrlQueryParams } from '../helper/utils'
 import * as webLifecycle from '../platform/patch/web/lifecycle'
 
 const webAppHooksMap = makeMap(webLifecycle.LIFECYCLE.APP_HOOKS)
@@ -25,17 +25,26 @@ export default function createApp (option, config = {}) {
   const builtInMixins = []
   const appData = {}
   if (__mpx_mode__ === 'web') {
+    const currentUrl = window.location.href
+    const urlQueryParams = getUrlQueryParams(currentUrl)
+    const options = {
+      path: currentUrl,
+      query: urlQueryParams,
+      scene: 0, // 有待确定
+      shareTicket: '', // 有待确定
+      referrerInfo: {}
+    }
     builtInMixins.push({
       created () {
         Object.assign(this, option.proto)
         Object.assign(this, appData)
-        this.$options.onLaunch && this.$options.onLaunch.call(this, {})
+        this.$options.onLaunch && this.$options.onLaunch.call(this, options)
         global.__mpxAppCbs = global.__mpxAppCbs || {
           show: [],
           hide: []
         }
         if (this.$options.onShow) {
-          this.$options.onShow.call(this, {})
+          this.$options.onShow.call(this, options)
           global.__mpxAppCbs.show.push(this.$options.onShow.bind(this))
         }
         if (this.$options.onHide) {
